requirements.txt
asgiref==3.8.1
celery==5.4.0
Django==4.2.19
Faker==35.2.0
fastapi==0.115.8
pydantic==2.10.6
python-dotenv==1.0.1
Requests==2.32.3
Eventlet==0.39.0
redis==5.2.1
psycopg2==2.9.10
uvicorn[standard]
Python-Multipart==0.0.20
